requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
python-dateutil==2.8.2
pyyaml==6.0.1
lxml==5.1.0
playwright==1.49.0
